Add database configurations

+++ TravelMore.Domain/Hotels/Hotel.cs
@@ -10,14 +10,14 @@ public class Hotel(Guid id) : Entity<Guid>(id)
     private readonly List<Booking> _bookings = [];
     public IReadOnlyCollection<Booking> Bookings => _bookings;
     public string Description { get; } = string.Empty;
-    public int MaxNumberOfGuests { get; set; }
+    public short MaxNumberOfGuests { get; set; }
     public Money Price { get; set; } = new(0);
     public int OwnerId { get; }
     public Host Owner { get; } = null!;
 
     public bool IsAvailable(BookingSchedule schedule) => AreAllBookingsOutsideSchedule(schedule);
 
-    public Money CalculateTotalPayment(int numberOfGuests) => new(Price.Amount * numberOfGuests);
+    public Money CalculateTotalPayment(short numberOfGuests) => new(Price.Amount * numberOfGuests);
 
     private bool AreAllBookingsOutsideSchedule(BookingSchedule schedule) => _bookings.All(booking => !booking.IsOverLap(schedule.From, schedule.To));
